pullrequest_20230113

+++ src/libs/zbxserver/anomalystl.c
@@ -19,10 +19,9 @@
 
 #include "anomalystl.h"
 
-#include "common.h"
+#include "zbxnum.h"
 #include "log.h"
 #include "zbxeval.h"
-#include "../zbxalgo/vectorimpl.h"
 
 ZBX_PTR_VECTOR_DECL(VV, zbx_vector_history_record_t *)
 ZBX_PTR_VECTOR_IMPL(VV, zbx_vector_history_record_t *)
@@ -317,7 +316,7 @@ static void	apply_loess_smoothing(const zbx_vector_history_record_t *y, int n, i
 		{
 			int	nsh;
 
-			nsh = (int)((length + 1) / 2);
+			nsh = (length + 1) / 2;
 			nleft = 1;
 			nright = length;
 
@@ -346,7 +345,7 @@ static void	apply_loess_smoothing(const zbx_vector_history_record_t *y, int n, i
 		{
 			int	nsh;
 
-			nsh = (int)((length + 1) / 2);
+			nsh = (length + 1) / 2;
 
 			for (i = 1; i < n + 1; i = i + newnj)
 			{
@@ -396,7 +395,7 @@ static void	apply_loess_smoothing(const zbx_vector_history_record_t *y, int n, i
 				ys->values[j].value.dbl = ys->values[i].value.dbl + (delta * (j - i));
 		}
 
-		k = (int)(((n - 1)/newnj) * newnj + 1);
+		k = ((n - 1)/newnj) * newnj + 1;
 
 		if (k != n)
 		{
